# lib/api-zod/src/generated/types/schoolInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SchoolInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
}