# lib/api-zod/src/generated/types/schoolInvitationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SchoolRole } from './schoolRole';

export interface SchoolInvitationInput {
  role: SchoolRole;
  /**
     * @minimum 1
     * @maximum 30
     */
  expiresInDays?: number;
}