# lib/api-zod/src/generated/types/schoolMemberRoleUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SchoolRole } from './schoolRole';

export interface SchoolMemberRoleUpdate {
  role: SchoolRole;
}