# lib/api-zod/src/generated/types/schoolMembership.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { School } from './school';
import type { SchoolMember } from './schoolMember';

export interface SchoolMembership {
  school: School;
  membership: SchoolMember;
}