# lib/api-zod/src/generated/types/schoolRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type SchoolRole = typeof SchoolRole[keyof typeof SchoolRole];


export const SchoolRole = {
  admin: 'admin',
  principal: 'principal',
  staff: 'staff',
} as const;